Get ID models automatically
Models classes and parameters now is got exclusively from init file;
Models types (APPLE, Delta, ...) now are checked and removed from models dict by searching the base classes of the model, which, in this case, is only InsertionDeviceModel;
Models parameters dict, which contains the parameters and default values of each specific model initialization, got from signature of the init function using inspect module.

getUndulatorName, Phase and isCorreted moved from init to data_dialog.

countIDnames method became deprecated, then was deleted from ProjectWidget.
